mv find_valid_samplers to mcmc._utilities
