+++ cuqi/experimental/mcmc/__init__.py
@@ -1,6 +1,6 @@
 """ Re-implementation of sampler module in a more object oriented way. """
 
-from ._sampler import SamplerNew, ProposalBasedSamplerNew, find_valid_samplers
+from ._sampler import SamplerNew, ProposalBasedSamplerNew
 from ._langevin_algorithm import ULANew, MALANew
 from ._mh import MHNew
 from ._pcn import PCNNew
@@ -12,3 +12,4 @@ from ._gibbs import HybridGibbsNew
 from ._conjugate import ConjugateNew
 from ._conjugate_approx import ConjugateApproxNew
 from ._direct import DirectNew
+from ._utilities import find_valid_samplers
